Untested: Should be possible to cache trips locally and view routes on map.

+++ UbiForsikring/app/src/main/java/sw10/ubiforsikring/Helpers/ServiceHelper.java
@@ -108,7 +108,7 @@ public class ServiceHelper {
 		return facts;
 	}
 
-	public static ArrayList<Fact> GetFactsForMap(int carid, int tripid){
+	public static ArrayList<Fact> GetFactsForMap(int carid, long tripid){
 		String response = "Empty response";
 
 		try {
